Moved "ActionsPageSteps" to "cypress/steps/pages"
