+++ cypress/e2e/editor-actions/configure-actions.spec.cy.ts
@@ -1,6 +1,6 @@
 import {YasqeSteps} from "../../steps/yasqe-steps";
 import {QueryStubs} from "../../stubs/query-stubs";
-import ActionsPageSteps from "../../steps/actions-page-steps";
+import ActionsPageSteps from "../../steps/pages/actions-page-steps";
 import {YasguiSteps} from "../../steps/yasgui-steps";
 
 describe('Configure editor actions', () => {
