+++ cypress/e2e/editor-actions/expand-results-over-sameas.spec.cy.ts
@@ -1,6 +1,6 @@
 import {YasqeSteps} from "../../steps/yasqe-steps";
 import {QueryStubDescription, QueryStubs} from "../../stubs/query-stubs";
-import ActionsPageSteps from "../../steps/actions-page-steps";
+import ActionsPageSteps from "../../steps/pages/actions-page-steps";
 
 describe('Expand results over sameAs', () => {
   beforeEach(() => {
