+++ cypress/e2e/editor-actions/include-inferred-statements.spec.cy.ts
@@ -1,6 +1,6 @@
 import {YasqeSteps} from "../../steps/yasqe-steps";
 import {QueryStubDescription, QueryStubs} from "../../stubs/query-stubs";
-import ActionsPageSteps from "../../steps/actions-page-steps";
+import ActionsPageSteps from "../../steps/pages/actions-page-steps";
 
 describe('Include inferred action', () => {
   beforeEach(() => {
