+++ cypress/e2e/saved-query/delete-saved-query.spec.cy.ts
@@ -1,6 +1,6 @@
 import {YasqeSteps} from "../../steps/yasqe-steps";
 import {QueryStubs} from "../../stubs/query-stubs";
-import ActionsPageSteps from "../../steps/actions-page-steps";
+import ActionsPageSteps from "../../steps/pages/actions-page-steps";
 import {ConfirmationDialogSteps} from "../../steps/confirmation-dialog.steps";
 
 describe('Delete saved query action', () => {
